Frontend move entityStyle, convertSchema and convertQueryResult to seperate files
